both are working
option 1 uses the TOF sensor as part of the safety system. in this configuration, the sensor is attached to the sliding element of the window, facing vertically downwards. In the event that rain is detected and the window starts closing, if the TOF sensor detects a distance of less than 10cm (placeholder), then it interrupts the window closing process. To go back to normal operation, the emergency button is to be pressed, which will cause the window to open fully again, before it checks for rain, and closes if it is raining.

option 2 uses the TOF sensor to tell the program the current location of the window (in milimetres away from the sensor). as the sensor is now placed facing horizontally to the sliding element of the window, it is now able to tell the program how far away the window is from it. option 2 uses this info to prevent the problem in option 1, which is that when set from auto mode to manual mode, and then back to auto mode, the window will jump to it's last recorded location when it was in auto mode, so if the window was moved when it was still in manual mode, changing to auto mode will cause the window to dangerously jump to another position, potentially hurting the user.
therefore, option 2 relies on the user hitting the emergency button instead.

currrently, option 1 also has the warning flash built-in, but not option 2 yet.